More debugging around client-side udp transactions.
And some cleanup along the way.

+++ src/iperf_api.c
@@ -774,7 +774,7 @@ iperf_on_connect(struct iperf_test *test)
             }
         }
         if (test->settings->rate)
-            iperf_printf(test, "      Target Bitrate: %llu\n", test->settings->rate);
+            iperf_printf(test, "      Target Bitrate: %llu\n", (unsigned long long)test->settings->rate);
     }
 }
